requirements.txt
fastapi
uvicorn
mysql-connector-python
jinja2
requests
python-multipart